Run the simulation engine

Run the simulation engine

RunEngine(
  trt_list,
  common_pt_inputs = NULL,
  unique_pt_inputs = NULL,
  input_list = NULL
)

Arguments

trt_list
A vector of the names of the interventions evaluated in the simulation

common_pt_inputs
A list of inputs that change across patients but are not affected by the intervention

unique_pt_inputs
A list of inputs that change across each intervention

input_list
A list of all other inputs: drc,drq,psa_bool,init_event_list,evt_react_list, uc_lists = list(util_ongoing_list,util_instant_list,util_cycle_list,cost_ongoing_list,cost_instant_list,cost_cycle_list), input_out,ipd,trt_list,simulation,npats,n_sim

Value

A data frame with the simulation results